moved middleware again; hopefully fixed the zod validation of files with the z.custom
