+++ src/server/uploadthing.ts
@@ -5,7 +5,7 @@ import { utapi } from "uploadthing/server";
 import {NextApiRequest, NextApiResponse} from "next";
 import {getAuth} from "@clerk/nextjs/server";
 import {prisma} from "~/server/db";
-import middleware from "../../middleware";
+import middleware from "../middleware";
 
 
 const f = createUploadthing();
@@ -17,6 +17,7 @@ export const ourFileRouter = {
     // Set permissions and file types for this FileRoute
     .middleware((req: NextApiRequest, res: NextApiResponse) => {
       // This code runs on your server before upload
+      getAuth(req)
       console.log(req.url)
 
       // If you throw, the user will not be able to upload
